Restore tasks file
